german_learning_game/llm/base.py
from abc import ABC, abstractmethod
from typing import Any, Dict, Optional
import yaml


class LLMProvider(ABC):
    def __init__(self, config_path: str):
        self.config = self._load_config(config_path)
        self.prompts = self._load_prompts()

    @abstractmethod
    async def generate_text(self, prompt: str, **kwargs) -> str:
        """Generate text using the LLM provider."""
        pass

    @abstractmethod
    async def analyze_text(self, text: str, instruction: str) -> Dict[str, Any]:
        """Analyze text according to given instructions."""
        pass

    def _load_config(self, config_path: str) -> Dict[str, Any]:
        """Load LLM provider configuration from YAML file."""
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)

    def _load_prompts(self) -> Dict[str, str]:
        """Load prompt templates from the prompts.yaml file."""
        with open(self.config["prompts_path"], 'r') as f:
            return yaml.safe_load(f)


class AnthropicProvider(LLMProvider):
    async def generate_text(self, prompt: str, **kwargs) -> str:
        # Implementation for Anthropic's Claude
        pass

    async def analyze_text(self, text: str, instruction: str) -> Dict[str, Any]:
        # Implementation for text analysis with Claude
        pass


class OpenAIProvider(LLMProvider):
    async def generate_text(self, prompt: str, **kwargs) -> str:
        # Implementation for OpenAI's GPT models
        pass

    async def analyze_text(self, text: str, instruction: str) -> Dict[str, Any]:
        # Implementation for text analysis with GPT
        pass


class LocalProvider(LLMProvider):
    async def generate_text(self, prompt: str, **kwargs) -> str:
        # Implementation for local LLM
        pass

    async def analyze_text(self, text: str, instruction: str) -> Dict[str, Any]:
        # Implementation for text analysis with local LLM
        pass