german_learning_game/pyproject.toml
[project]
name = "german_learning_game"
version = "0.1.0"
description = "An interactive German language learning game with AI agents"
authors = [
    {name = "OpenHands", email = "openhands@all-hands.dev"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0",
    "smol-dev-python>=0.1.0",
    "pydantic>=2.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3